Add a flag to the SimpleMessageListenerContainer to control if a message must be deleted or not when an exception occurs in the handler method.

+++ spring-cloud-aws-messaging/src/main/java/org/springframework/cloud/aws/messaging/listener/SimpleMessageListenerContainer.java
@@ -44,6 +44,7 @@ public class SimpleMessageListenerContainer extends AbstractMessageListenerConta
 
 	private volatile CountDownLatch stopLatch;
 	private boolean defaultTaskExecutor;
+	private boolean deleteMessageOnExceptionHandling = true;
 
 	protected TaskExecutor getTaskExecutor() {
 		return this.taskExecutor;
@@ -51,6 +52,14 @@ public class SimpleMessageListenerContainer extends AbstractMessageListenerConta
 
 	public void setTaskExecutor(TaskExecutor taskExecutor) {
 		this.taskExecutor = taskExecutor;
+	}
+
+	public boolean isDeleteMessageOnExceptionHandling() {
+		return this.deleteMessageOnExceptionHandling;
+	}
+
+	public void setDeleteMessageOnExceptionHandling(boolean deleteMessageOnExceptionHandling) {
+		this.deleteMessageOnExceptionHandling = deleteMessageOnExceptionHandling;
 	}
 
 	@Override
@@ -181,7 +190,7 @@ public class SimpleMessageListenerContainer extends AbstractMessageListenerConta
 				executeMessage(queueMessage);
 				getAmazonSqs().deleteMessageAsync(new DeleteMessageRequest(this.queueUrl, receiptHandle));
 			} catch (MessagingException e) {
-				if (!this.hasRedrivePolicy) {
+				if (!this.hasRedrivePolicy && SimpleMessageListenerContainer.this.isDeleteMessageOnExceptionHandling()) {
 					getAmazonSqs().deleteMessageAsync(new DeleteMessageRequest(this.queueUrl, receiptHandle));
 				}
